requirements.txt
binance-connector==3.6.0
h5py==3.10.0
numpy==1.26.4
torch==2.2.1